DetailView Fragment
Added a detailed view fragment which is opened when the user taps on an
assignment. The fragment contains the title, description, class and due
date as well as a delete functionality. This was previously accomplished
using a dialog fragment, however the detail view fragment provides more
functionality and design flexibility. Next I will focus on adding
transitions as well as notifications.